Move FixedLengthArray to its own file
