+++ spec/metric_scaler_spec.rb
@@ -124,24 +124,5 @@ describe TopologicalInventory::Orchestrator::MetricScaler do
         watcher.scale_to_desired_replicas
       end
     end
-
-    describe TopologicalInventory::Orchestrator::MetricScaler::Watcher::FixedLengthArray do
-      subject { described_class.new(5) }
-
-      context "#average" do
-        it "with fewer inserts than the max size" do
-          2.times { subject << 2 }
-
-          expect(subject.average).to eq(2)
-        end
-
-        it "with more inserts than the max size" do
-          5.times { subject << 2 }
-          5.times { subject << 4 }
-
-          expect(subject.average).to eq(4)
-        end
-      end
-    end
   end
 end
